Changed player_iplong() to return an unsigned long
git-svn-id: http://hopmod.googlecode.com/svn/trunk@1633 d7818858-e1b4-11dd-a738-5fe2e3cfbcc5

+++ src/hopmod/extapi.hpp
@@ -64,7 +64,7 @@ namespace server
     int player_ping_update(int);
     int player_lag(int);
     const char * player_ip(int);
-    int player_iplong(int);
+    unsigned long player_iplong(int);
     const char * player_status(int);
     int player_status_code(int);
     int player_frags(int);
